Extract logic to parse bad request error messages into a method of an Error Parser class
The parsing of bad request error messages is being used across different modals, so we are extracting the logic into a method.
